+++ src/software/cpu/arch2d_neon.c
@@ -174,7 +174,7 @@ _batch_imadd_neon(int32_ptr dst, const_int32_ptr src, size_t num, float v, float
    size_t i, size, step;
 
    if (!num || (v == 0.0f && vstep == 0.0f)) return;
-   if (fabs(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
+   if (fabsf(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
       _batch_iadd_neon(dst, src, num);
       return;
    }
@@ -285,7 +285,7 @@ _batch_fmadd_neon(float32_ptr dst, const_float32_ptr src, size_t num, float v, f
    size_t i, step;
 
    if (!num || (v == 0.0f && vstep == 0.0f)) return;
-   if (fabs(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
+   if (fabsf(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
       _batch_fadd_neon(dst, src, num);
       return;
    }
