+++ src/software/cpu/arch2d_sse2.c
@@ -500,7 +500,7 @@ _batch_imadd_sse2(int32_ptr dst, const_int32_ptr src, size_t num, float v, float
    size_t i, step, dtmp, stmp;
 
    if (!num || (v == 0.0f && vstep == 0.0f)) return;
-   if (fabs(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
+   if (fabsf(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
       _batch_iadd_sse2(dst, src, num);
       return;
    }
@@ -716,7 +716,7 @@ _batch_fmadd_sse2(float32_ptr dst, const_float32_ptr src, size_t num, float v, f
    size_t i, step, dtmp, stmp;
 
    if (!num || (v == 0.0f && vstep == 0.0f)) return;
-   if (fabs(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
+   if (fabsf(v - 1.0f) < GMATH_128DB && vstep == 0.0f) {
       _batch_fadd_sse2(dst, src, num);
       return;
    }
